PET-127 refactor : common 패키지 제거 및 내부에 있던 패키지 외부로 이동
