+++ Domain-Module/Auth-Module/Auth-Application/src/main/java/com/pawith/authapplication/config/FeignClientConfig.java
@@ -1,4 +1,4 @@
-package com.pawith.authapplication.common.config;
+package com.pawith.authapplication.config;
 
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Configuration;
