+++ Domain-Module/Auth-Module/Auth-Presentation/src/main/java/com/pawith/authpresentation/common/security/filter/JWTAuthenticationFilter.java
@@ -3,8 +3,8 @@ package com.pawith.authpresentation.common.security.filter;
 import com.pawith.authapplication.service.JWTExtractEmailUseCase;
 import com.pawith.authapplication.service.JWTExtractTokenUseCase;
 import com.pawith.authapplication.service.JWTVerifyUseCase;
-import com.pawith.authapplication.common.consts.AuthConsts;
-import com.pawith.authapplication.common.consts.IgnoredPathConsts;
+import com.pawith.authapplication.consts.AuthConsts;
+import com.pawith.authapplication.consts.IgnoredPathConsts;
 import com.pawith.authpresentation.common.security.JWTAuthenticationToken;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
